move main procedure to function `util.alphaevolve_minimization`
